Add market stock buying and price scaling

+++ src/backend/spacetimedb/module_bindings/set_npc_market_item_base_price_reducer.ts
@@ -12,5 +12,5 @@ import {
 
 export default {
   itemKey: __t.string(),
-  basePriceGold: __t.u64(),
+  basePriceGold: __t.f64(),
 };
